requirements.txt
aiofiles==24.1.0
blinker==1.5
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
google-ai-generativelanguage==0.4.0
google-api-core==2.24.2
google-auth==2.39.0
google-generativeai==0.3.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
h2==4.2.0
hpack==4.1.0
hypercorn==0.17.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
priority==2.0.0
proto-plus==1.26.1
protobuf==4.25.6
pyasn1==0.6.1
pyasn1-modules==0.4.2
quart==0.18.4
requests==2.32.3
rsa==4.9
tqdm==4.67.1
urllib3==2.4.0
werkzeug==2.3.8
wsproto==1.2.0
youtube-transcript-api==0.6.1
